Envoy full and boostrap config added, a little bit of refactoring
